backend/requirements.txt
Django>=5.1
djangorestframework>=3.16
djangorestframework-simplejwt>=5.4
django-cors-headers>=4.4
dj-database-url>=2.2
psycopg2-binary>=2.9
channels>=4.1
channels-redis>=4.2
daphne>=4.1
whitenoise>=6.8
